Compare two files as specify location of others
New function to load two files in two windows where they can be compared
since they sit next to each other. This can either be below or to the
right.

All functions now allow you to specify if the new side window is place
on the right side or below.