Add an "advanced" keybinding mode
While the Space bar was a great idea for some presentations, it doesn't
lend itself for doing actual work during each step, so now we can call
the 'demo-it-start' function with optional 't' to change from the space
bar to rebinding the F1 key.

Also, added a demo-it-typewriter function for simulating the typing of
text into a buffer.