Make org-mode presentations more presentable
When calling `demo-it-presentation` with an `org-mode` file, the source
code blocks' `begin` and `end` header lines are *hidden* to make the
source code stand out and not confuse people looking at a presentation.

This also sets the `org-hide-emphasis-markers` in order to remove the
visual distractions of the bold, italics and other characters.

All these changes are restored when the `demo-it-presentation-quit` is
called.