Removal of org-block-background
With the org-mode changes from `org-block-background` to just
`org-block`, I had to change the presentation mode. The biggest change
is that if you have syntax highlighting for a code block, and variable
pitch for the rest of the presentation, the variable pitch will carry on
into the code blocks.

So, variable pitch is now an optional parameter to the
`demo-it-presentation' until I can figure out a way to make a distinction.